backend/requirements.txt
astroid==2.2.3
autopep8==1.4.3
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
isort==4.3.10
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2018.9
rope==0.12.0
six==1.12.0
typed-ast==1.2.0
wrapt==1.11.1
